requirements.txt
pillow
pytils
django-ckeditor
